fix: update ExecutionEndpoint signature for float multiplier
Mirrors the type change in BundlerManager/InitData so the constructor
type matches the value passed through from cli_manager.

+++ voltaire_bundler/execution_endpoint.py
@@ -86,7 +86,7 @@ class ExecutionEndpoint(Endpoint):
         is_eip7702: bool,
         min_stake: int,
         min_unstake_delay: int,
-        bundle_gas_estimation_multiplier: int,
+        bundle_gas_estimation_multiplier: float,
     ):
         super().__init__("bundler_endpoint")
         self.ethereum_node_urls = ethereum_node_urls
